feat: 支持展示未同步风险和事件的占位状态 --story=129054781

+++ src/backend/services/web/risk/handlers/risk.py
@@ -224,10 +224,10 @@ class RiskHandler:
 
         # 存在则更新结束时间, 风险事件描述
         if risk:
-            last_end_time = int(event["event_time"] / 1000)
+            last_end_time = event["event_time"] / 1000
             logger.info("[UpdateRisk] Risk exists. risk_id=%s; last_end_time=%s", risk.risk_id, last_end_time)
             # 只在事件的时间更新的时候存储
-            if int(risk.event_end_time.timestamp()) < last_end_time:
+            if risk.event_end_time.timestamp() < last_end_time:
                 risk.event_end_time = datetime.datetime.fromtimestamp(last_end_time)
                 risk.save(update_fields=["event_end_time"])
             if event.get("event_content") and risk.event_content != event["event_content"]:
